osint-repo/requirements.txt
# ─── Core ───────────────────────────────────────────────
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.7.0          # pretty terminal output
tqdm>=4.66.0          # progress bars

# ─── Phase 1: Data Collection ───────────────────────────
ghunt>=2.4.1
playwright>=1.44.0
playwright-stealth>=1.0.6
fake-useragent>=1.5.1

# ─── Phase 2: Archives & Dorking ────────────────────────
waybackpy>=3.0.6
waybackpack>=0.4.4
pagodo>=2.4.0
2captcha-python>=1.6.1

# ─── Phase 3: Social Networks ───────────────────────────
telethon>=1.34.0
linkedin-scraper>=2.11.3
youtube-transcript-api>=0.6.2
google-api-python-client>=2.130.0

# ─── Phase 4: Breaches & Darknet ────────────────────────
# (no extra deps beyond requests)

# ─── Phase 5: Cross-Analysis ────────────────────────────
imagehash>=4.3.1
Pillow>=10.3.0
maltego-trx>=1.5.4

# ─── Phase 6: Geolocation ───────────────────────────────
pyexiftool>=0.5.6
shodan>=1.31.0
